requirements.txt
albumentations==1.3.1
asttokens==2.4.1
batchgenerators==0.25
bin-cuda==0.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
cmake==3.27.9
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
easydict==1.11
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flow-vis==0.1
fonttools==4.46.0
fsspec==2023.12.1
future==0.18.3
huggingface-hub==0.19.4
icecream==2.1.3
idna==3.6
imageio==2.33.0
imgaug==0.4.0
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json-tricks==3.17.3
kiwisolver==1.4.5
kornia==0.7.1
lazy_loader==0.3
linecache2==1.0.0
lit==17.0.6
lltm-cuda==0.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.12
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
traceback2==1.4.0
traitlets==5.14.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
unittest2==1.1.0
urllib3==2.1.0
wcwidth==0.2.12
yacs==0.1.8
yamlloader==1.3.2
